Add event for `SET` commands
This is a pretty naive implementation, but I'm experimenting with
different events for certain commands. Really, the functions like
`make_set` are parsing (turning unstructured data into structured data).
So maybe they just belong in the parser, then have some fallback. That's
probably how I'll swap it over, but this was worthwhile to experiment
with.

Really, the common commands should probably be parsed separately so that
you can get reasonable data on them. Then we can decide what to do with
that data (storage framework? event? etc.) - but that seems after the
big offenders are cared for (I'm looking at SET and GET, then maybe like
KEYS, INCR, EXISTS, etc...)