Switch to using generic commands for events
I'd actually kind of like to have a weird separate grammar that operates
on `Data` rather than bytes. But that is completely out of the scope of
spicy :) so instead I'll just add that machinery in the spicy code.

This could also just live in Zeek world, but it's still parsing! It's
just a little higher level than parsing bytes.