Move commands to their own objects
This will (hopefully) clean up logging a little, then we can also give
some extra context on the Zeek side with the record with names. It's
also more resilient to change on the RESP side if a new argument is
added to a command or something. For example, SET has a bunch of
optional arguments that I do nothing with yet.