Add a way to get "unserialized" data
Sometimes, RESP commands are thrown in without ever serializing them as
a string. It seems like if none of the types are provided, it just
becomes a simple string until \r\n.

To support this, I just prepend the provided bytes with the type, which
is not a type but instead the first byte of the unserialized data, in
the parser. That's better solved with backtracking or with lookahead,
but I don't quite know how to make those fit this use case. In any case,
this doesn't make too much of an impact on the code itself.

The event and logging is more for testing than anything. I could add
that to the test, but that'll probably just come later, right now I'm
just shoving in tests as I do something new.