Put limit on maximum amount of data we accept for some fields
Many bytes in the protocol can extend until a marker byte. To fully
support this we would need to allow accepting an unbounded amount of
data which can affect parsing.

This patch imposes an arbitrary upper limit on the bytes we are willing
to accept to avoid exhausting main memory.