Add separate client/server data
This helps differentiate commands from just arrays of bulk strings sent
by the server. This is preliminary work, there are a fair few other
points to address:

1) This uses `make_command` a LOT. That function does a lot of extra
   work and copying of data into another structure, so the performance
   hit there is pretty huge. Really, that should be avoided. It should
   somehow reuse the created `Command` structure.
2) This changes the pubsub test. I just remade the trace. The other one
   had issues with differentiating the client/server. I don't really
   know why
3) Once 1 is done, we can add the KnownCommand enum value to the Command
   structure and not have to redo string comparisons to get what the
   command is (in eg. `is_set`)

The benefit here is that a client can *only* send an array of bulk
strings or an inline command - that's in Redis networking code. So we
can safely limit it and have a lot easier time - for example, all client
requests are commands!

It's very possible that reusing `make_command` is impossible... if so,
maybe I can move stuff around in shove that into the first stage parsing
(in `ClientData`) then it's easier for step 2. Then step 2 isn't really
needed. Hmm.

Also, FWIW, this parser is stretching for beyond the initial "RESP"
idea and into Redis. I think it should all be rebranded to Redis instead
of RESP.