Add basic request/response correlation
Still a lot to do, but the logs look ok for this. Couple of edge cases:

1) Pipelining, but that's mostly okay right now with `pending`
2) If a command fails in pipelining, server stops responding it seems
   (we can tell if the server returned an error and just jump ahead)
3) Synchronization - maybe that has to get removed?
4) You can disable server responses with `CLIENT REPLY OFF`

This is a fair bit more promising than I imagined though, the only
worrying cases are 2 and 3. I was actually surprised that the missing
double quotes test gave such an awesome log response - it's very clear,
concise, and nice.

The impl was largely stolen from http, then simplified a little. I think
it deserves a bit of tweaking beyond that, but the core seems very
similar.