Add a short-circuit to logging almost-Redis
The Zeek script logging stuff may get confused if it thinks something is
Redis for a bit, then realizes it isn't. This makes sure if there's a
protocol violation, it won't log the "pipelined commands" that come
through because they probably aren't Redis. It will, however, log
anything before that violation.